==> 03_data_analysis.Rmd <==
---
title: "Epi 3 Data Analysis"
author: "Qiheng(Michael) Yan"
date: "2023-11-21"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
require("knitr")
opts_knit$set(root.dir = here::here('data'))
```

```{r, message=FALSE}
library(tidyverse)
library(stats)
library(ggplot2)
library(expss)
set.seed(1028)
```

# Epi 3 Data Analysis

## Import data
```{r}
# Import the dataset
complete_data<-readRDS("combined_data.rds")
subset_1168_data<-readRDS("subset_1168.rds")
#write.csv(complete_data, "complete_data.csv", row.names=FALSE)
#write_labelled_csv(complete_data, "complete_data_labeled.csv", row.names=FALSE)
#write.csv(subset_1168_data, "subset_1168_data.csv", row.names=FALSE)
#write_labelled_csv(subset_1168_data, "subset_1168_data_labeled.csv", row.names=FALSE)
```

This analysis follows the inclusion/exclusion criteria and uses a subset of 1168 subjects (using subset_1168_data)

## Bivariate Analysis
In bivariate analysis, we'll use Chi-squared tests for categorical variables like health insurance status, race/ethnicity, gender, and a T-test for continuous variables like age.
```{r}
# Chi-squared test for medication adherence and family income
chisq_income <- chisq.test(table(subset_1168_data$adherence, subset_1168_data$income_cat))

# Chi-squared test for medication adherence and total insurance status
chisq_insurance <- chisq.test(table(subset_1168_data$adherence, subset_1168_data$ins_classif))

# Chi-squared test for medication adherence and race
chisq_race <- chisq.test(table(subset_1168_data$adherence, subset_1168_data$race_6cat))

# Chi-squared test for medication adherence and sex
chisq_gender <- chisq.test(table(subset_1168_data$adherence, subset_1168_data$sex))

# Chi-squared test for medication adherence and education level
chisq_edu <- chisq.test(table(subset_1168_data$adherence, subset_1168_data$educ_level))

# Chi-squared test for medication adherence and marital status
chisq_marital <- chisq.test(table(subset_1168_data$adherence, subset_1168_data$marital_status))

# Chi-squared test for medication adherence and alcohol use
chisq_alcohol <- chisq.test(table(subset_1168_data$adherence, subset_1168_data$alc_heavy))

# Chi-squared test for medication adherence and tobacco use
chisq_tobacco <- chisq.test(table(subset_1168_data$adherence, subset_1168_data$smoking_hx))

# T-test for medication adherence and age
t_test_age <- t.test(subset_1168_data$age ~ subset_1168_data$adherence)

# Print the results
print(chisq_income)
print(chisq_insurance)
print(chisq_race)
print(chisq_gender)
print(chisq_edu)
print(chisq_marital)
print(chisq_alcohol)
print(chisq_tobacco)
print(t_test_age)
```

The results from the bivariate analyses provide valuable insights into the relationships between adherence to prescribed cholesterol medication and various factors such as family income, health insurance status, race/ethnicity, gender, marital status, alcohol use, tobacco use, and age. 

1. **Family Income vs. Adherence**:

**Chi-squared test result**: 
$$X^2=7.6725, df=2, p-value=0.02157$$
**Interpretation**: There is a statistically significant association between family income and medication adherence. The low p-value suggests that differences in medication adherence across different income categories are not likely due to chance.

2. **Health Insurance Status vs. Adherence**:

**Chi-squared test result**: 
$$X^2=32.62, df=8, p-value=7.209 \times 10^{-5}$$
**Interpretation**: There is a strong statistically significant association between insurance status and medication adherence. The very low p-value indicates that the observed differences in adherence across various insurance classifications are unlikely to be due to random variation.

3. **Race/Ethnicity vs. Adherence**:

**Chi-squared test result**: 
$$X^2=9.8825, df=5, p-value=0.07863$$
**Interpretation**: The association between race and medication adherence is not statistically significant at the conventional alpha level of 0.05. This suggests that differences in adherence across different racial categories may be due to chance.

4. **Gender vs. Adherence**:

**Chi-squared test result**: 
$$X^2=4.1148, df=1, p-value=0.04251$$
**Interpretation**: There is a statistically significant association between gender and medication adherence, with the p-value indicating that these differences are unlikely to be due to random chance.

5. **Education Level vs. Adherence**:

**Chi-squared test result**: 
$$X^2=2.2859, df=4, p-value=0.6833$$
**Interpretation**: The p-value of 0.6833 is well above the conventional alpha level of 0.05, which suggests that there is no statistically significant association between education level and adherence to cholesterol medication in this sample.

6. **Marital Status vs. Adherence**:

**Chi-squared test result**: 
$$X^2=14.11, df=2, p-value=0.0008633$$
**Interpretation**: This suggests that marital status is an important factor in adherence.

7. **Alcohol Use vs. Adherence**:

**Chi-squared test result**: 
$$X^2=3.7807, df=1, p-value=0.05185$$
**Interpretation**: The association between heavy alcohol use and medication adherence is on the borderline of statistical significance, indicating a potential relationship that may require further exploration.

8. **Tobacco use vs. Adherence**:

**Chi-squared test result**: 
$$X^2=14.247, df=2, p-value=0.0008058$$
**Interpretation**: The association between heavy alcohol use and medication adherence is on the borderline of statistical significance, indicating a potential relationship that may require further exploration.

9. **Age vs. Adherence**:

**T-test result**:
$$t=-8.921, df=692.18, p-value=2.2 \times 10^{-16}$$
**Interpretation**: There is a significant association between tobacco use and medication adherence, suggesting that smoking status is an important factor.

**Note on the Warning**: The warning regarding the Chi-squared approximation may be due to small expected frequencies in some cells of the contingency tables. This is common when some categories have a low number of observations. In such cases, alternative tests like Fisher's Exact Test for small sample sizes might be more appropriate, especially for the race and gender analyses.

### Bivariate Analysis Interpretation
The bivariate analysis indicates several noteworthy relationships between demographic, socio-economic, and behavioral factors and medication adherence. Income level, insurance status, marital status, and tobacco use all show statistically significant correlations with adherence to cholesterol medication, suggesting these factors may play a critical role in whether individuals follow their prescribed medication regimen. Additionally, age remains a strong factor, with older individuals more likely to adhere to their medication. Although the associations with race/ethnicity and alcohol use did not reach statistical significance, they may still be relevant in the broader context of medication adherence and are worth exploring further. These bivariate findings lay the groundwork for more comprehensive analyses that can untangle the complex interplay of these variables and inform targeted strategies for improving adherence.

## Bivariate Analysis Visualization
```{r}
# Ensure NA is a factor level and label it as "No Response"
subset_1168_data$adherence_factor <- factor(subset_1168_data$adherence, levels = c(FALSE, TRUE))
subset_1168_data$adherence_factor <- addNA(subset_1168_data$adherence_factor)
levels(subset_1168_data$adherence_factor)[is.na(levels(subset_1168_data$adherence_factor))] <- "No Response"

# Define new colors for the bars, including NA
new_colors <- c("TRUE" = "#1b9e77", "FALSE" = "#d95f02", "No Response" = "#4D4D4D")

# Create the plots, making sure to use scale_fill_manual to include NA values
# and set the axis titles correctly after coord_flip()

# Income Category vs Adherence
ggplot(subset_1168_data, aes(x = income_cat, fill = adherence_factor)) +
  geom_bar() +
  scale_fill_manual(values = new_colors) +
  labs(title = "Adherence to Cholesterol Medication by Income Category",
       y = "Count",
       x = "Income Category") +
  theme_minimal() +
  coord_flip() +
  theme(legend.title = element_blank())

# Health Insurance Classification vs Adherence
ggplot(subset_1168_data, aes(x = ins_classif, fill = adherence_factor)) +
  geom_bar() +
  scale_fill_manual(values = new_colors) +
  labs(title = "Adherence to Cholesterol Medication by Health Insurance Status",
       y = "Count",
       x = "Health Insurance Classification") +
  theme_minimal() +
  coord_flip() +
  theme(legend.title = element_blank())

# Race/Ethnicity vs Adherence
ggplot(subset_1168_data, aes(x = race_6cat, fill = adherence_factor)) +
  geom_bar() +
  scale_fill_manual(values = new_colors) +
  labs(title = "Adherence to Cholesterol Medication by Race/Ethnicity",
       y = "Count",
       x = "Race/Ethnicity") +
  theme_minimal() +
  coord_flip() +
  theme(legend.title = element_blank())

# Gender vs Adherence
ggplot(subset_1168_data, aes(x = sex, fill = adherence_factor)) +
  geom_bar() +
  scale_fill_manual(values = new_colors) +
  labs(title = "Adherence to Cholesterol Medication by Gender",
       y = "Count",
       x = "Gender") +
  theme_minimal() +
  coord_flip() +
  theme(legend.title = element_blank())

# Education Level vs Adherence
ggplot(subset_1168_data, aes(x = educ_level, fill = adherence_factor)) +
  geom_bar() +
  scale_fill_manual(values = new_colors) +
  labs(title = "Adherence to Cholesterol Medication by Education Level",
       y = "Count",
       x = "Education Level") +
  theme_minimal() +
  coord_flip() +
  theme(legend.title = element_blank())

# Marital Status vs Adherence
ggplot(subset_1168_data, aes(x = marital_status, fill = adherence_factor)) +
  geom_bar() +
  scale_fill_manual(values = new_colors) +
  labs(title = "Adherence to Cholesterol Medication by Marital Status",
       y = "Count",
       x = "Marital Status") +
  theme_minimal() +
  coord_flip() +
  theme(legend.title = element_blank())

# Alcohol Use vs Adherence
ggplot(subset_1168_data, aes(x = alc_heavy, fill = adherence_factor)) +
  geom_bar() +
  scale_fill_manual(values = new_colors) +
  labs(title = "Adherence to Cholesterol Medication by Alcohol Use",
       y = "Count",
       x = "Alcohol Use") +
  theme_minimal() +
  coord_flip() +
  theme(legend.title = element_blank())

# Tobacco Use vs Adherence
ggplot(subset_1168_data, aes(x = smoking_hx, fill = adherence_factor)) +
  geom_bar() +
  scale_fill_manual(values = new_colors) +
  labs(title = "Adherence to Cholesterol Medication by Tobacco Use",
       y = "Count",
       x = "Tobacco Use") +
  theme_minimal() +
  coord_flip() +
  theme(legend.title = element_blank())

# Age vs Adherence Box Plot
ggplot(subset_1168_data, aes(x = adherence_factor, y = age, fill = adherence_factor)) +
  geom_boxplot() +
  scale_fill_manual(values = new_colors) +
  labs(title = "Age Distribution by Cholesterol Medication Adherence",
       x = "Adherence to Medication",
       y = "Age") +
  theme_minimal() +
  theme(legend.title = element_blank())
```

The resulting plots reveal several patterns related to adherence to cholesterol medication among adults:

1. **Adherence to Cholesterol Medication by Income Category**: The bar chart displays adherence to cholesterol medication across different income categories. The group with an income greater than 185% of the Federal Poverty Level (FPL) shows a larger proportion of individuals adhering to medication compared to the other groups, which aligns with the statistical significance found in the chi-squared test.

2. **Adherence to Cholesterol Medication by Health Insurance Status**:
This chart demonstrates that individuals with government health insurance and drug coverage show higher adherence compared to those without drug coverage or with unknown coverage. The significant chi-squared test result is visually corroborated here, showing the importance of insurance and drug coverage on medication adherence.

3. **Adherence to Cholesterol Medication by Race/Ethnicity**:
The distribution across race/ethnicity groups shows that Non-Hispanic White individuals have the highest count of adherence, followed by Non-Hispanic Black individuals. The chi-squared test result was not statistically significant, suggesting that the observed differences might be due to chance. However, the plot still shows notable differences in adherence rates that may warrant further investigation.

4. **Adherence to Cholesterol Medication by Gender**:
The chart indicates that males have a slightly higher count of non-adherence than females. The chi-squared test for gender was significant, suggesting a potential difference in adherence behavior between genders.

5. **Adherence to Cholesterol Medication by Education Level**: This bar chart illustrates the relationship between education level and adherence to cholesterol medication. It shows that individuals with some college or an associate's degree tend to adhere more to their cholesterol medication regimen compared to other educational levels. However, this trend does not present a clear gradient with increasing education, as those with a high school diploma or GED show a lower adherence rate, which is slightly less than those who didn't complete high school. Meanwhile, individuals with college degrees or higher show a moderate adherence rate. The chi-squared test did not find a statistically significant association between education level and adherence, suggesting that factors other than education level may play a more influential role in determining medication adherence.

6. **Adherence to Cholesterol Medication by Marital Status**: The visualization suggests that marital status may play a role in adherence to cholesterol medication. Individuals who are married or living with a partner show a higher adherence rate compared to those who have never been married or are widowed/divorced/separated. This is consistent with the chi-squared test indicating a significant association between marital status and medication adherence.

7. **Adherence to Cholesterol Medication by Alcohol Use**: The bar chart suggests that heavy alcohol users have a slightly lower adherence to cholesterol medication than those who do not engage in heavy drinking, as indicated by the larger proportion of non-adherence in the 'TRUE' category for heavy drinking. However, this association is not statistically significant at the 0.05 level, with a p-value just above the threshold.

8. **Adherence to Cholesterol Medication by Tobacco Use**: The chart illustrates adherence to cholesterol medication among different categories of tobacco users. Individuals who have never smoked show a higher proportion of adherence to medication, followed closely by past smokers, while current smokers have the lowest proportion of adherence. The chi-squared test result supports a significant association between tobacco use and medication adherence.

9. **Age Distribution by Cholesterol Medication Adherence**: The box plot shows a clear difference in the age distribution between those who adhere and those who do not, with the median age of adherent individuals being higher. This is consistent with the t-test results, which indicated a significant difference in age between the two groups.

### Bivariate Analysis Visualization Interpretation
Our bivariate analysis visualizations comprehensively encapsulate factors like socioeconomic status, insurance coverage, demographic characteristics, and personal behaviors, including marital status, alcohol, and tobacco use. These visual representations are crucial for a quick and clear understanding of the data and for effectively sharing our findings.

Through these visualizations, we see that marital status continues to show a strong link with medication adherence, suggesting that being married or having a partner may contribute to better health practices. Tobacco use also appears to be an influential factor; individuals who have never smoked tend to adhere more to medication regimens, potentially indicative of overall healthier lifestyle choices. Alcohol use does not show a statistically significant association with adherence, but the pattern observed could be informative in a broader health context.

The bivariate analysis also sheds light on the role of education in medication adherence. The visualizations do not reveal a clear-cut trend across education levels, suggesting that higher education does not necessarily equate to higher adherence. This could imply that education level alone is not a straightforward predictor of adherence, and that it must be considered alongside other factors that may exert a more complex influence.

In summary, these visualizations underscore the intricate web of factors that interplay in the realm of medication adherence. They stress the importance of considering a range of variables, including demographic, socioeconomic, behavioral, and educational, to gain a full understanding. This comprehensive approach is vital for crafting effective interventions and policies to improve adherence rates and tackle health disparities. The next step in our analysis, a multivariate approach, will be critical to unravel the combined impact of these diverse factors on adherence to cholesterol medication.



## Multivariate Analysis
Next, we'll conduct a logistic regression analysis to assess the relationship between family income and medication adherence, controlling for confounders.
```{r}
# logistic regression model with all variables
lr_model <- glm(adherence ~ income_cat + ins_classif + race_6cat + sex + age + 
                        educ_level + marital_status + alc_heavy + smoking_hx, 
                        family = binomial(link = "logit"), 
                        data = subset_1168_data)

# Summary of the model
summary(lr_model)
```

The logistic regression model provides an estimate of the association between various predictor variables and the likelihood of adherence to cholesterol medication. Here's an interpretation of the key findings from the model output:

1. **Income Category**: The coefficients for income_cat.L (linear term) and income_cat.Q (quadratic term) are not statistically significant (p > 0.05), suggesting no clear association between income category and medication adherence based on the available data.

2. **Insurance Classification**: None of the insurance classification terms (linear through to the eighth polynomial term) are statistically significant, indicating that within this model, insurance classification does not have a statistically significant association with medication adherence.

3. **Race/Ethnicity**: The linear term race_6cat.L is significant (p = 0.021), suggesting that there may be differences in medication adherence across different racial/ethnic categories. However, higher-order terms are not significant, indicating that the relationship is not straightforward and may not follow a clear pattern across all categories.

4. **Sex**: The term sex.L (representing sex of the individual) is not significant, implying that sex does not have a statistically significant impact on medication adherence in this model.

5. **Age**: Age is significant (p < 0.0001), indicating a positive association with adherence to cholesterol medication. This suggests that as age increases, the likelihood of adhering to medication also increases.

6. **Education Level**: Education level terms are not statistically significant, suggesting no clear pattern or association between different levels of education and medication adherence.

7. **Marital Status**: The terms for marital status are not statistically significant, indicating that marital status, as categorized in this model, does not have a significant impact on medication adherence.

8. **Alcohol Use**: The term for heavy alcohol use (alc_heavyTRUE) is not statistically significant, suggesting that heavy alcohol use does not have a significant impact on medication adherence in this model.

9. **Tobacco Use**: The linear term for tobacco use (smoking_hx.L) is significant (p = 0.020), indicating that there is a significant association between tobacco use and medication adherence. The negative coefficient suggests that individuals with a history of tobacco use are less likely to adhere to their medication. The quadratic term (smoking_hx.Q) approaches significance (p = 0.0787), indicating a possible non-linear relationship between tobacco use history and medication adherence.

Overall, the model indicates that age and certain aspects of race/ethnicity and tobacco use history are associated with medication adherence, while other factors such as income, insurance classification, sex, education level, marital status, and heavy alcohol use are not significantly associated in this analysis. It's important to note that the lack of significance for some variables does not necessarily mean that they have no effect on medication adherence but rather that any effect is not detectable with this model and dataset. Additional analysis, perhaps with different modeling strategies or additional data, may be required to fully understand these relationships.

```{r}
# Calculate Odds Ratios and 95% Confidence Intervals
or <- exp(coef(lr_model))
# Wald Confidence Intervals and p-values
se <- sqrt(diag(vcov(lr_model)))
wald_ci_lower <- exp(coef(lr_model) - 1.96 * se)
wald_ci_upper <- exp(coef(lr_model) + 1.96 * se)
p_values <- summary(lr_model)$coefficients[, "Pr(>|z|)"]

# Create a data frame to nicely format the results
results <- data.frame(
  OR = exp(coef(lr_model)),
  LowerCI = wald_ci_lower,
  UpperCI = wald_ci_upper,
  PValue = p_values
)

# View the results
print(results)
```

```{r}
####### Uncomment this chunk to generate the multivariate results table #######
# # Load the required packages
# library(knitr)
# library(kableExtra)
# 
# # Create a nice looking table to present the above ORs, CIs, and p-values
# nice_table <- kable(results,
#                     format = "html",  # Use "latex" for PDF output or "pipe" for Markdown or "html" for HTML
#                     digits = 3,       # Number of decimal places
#                     align = 'c',      # Center align the columns
#                     caption = "Multivariate Analysis of Factors Associated with Adherence to Cholesterol Medication") %>%
#   kable_styling(bootstrap_options = c("striped", "hover", "condensed"),
#                 full_width = F,
#                 position = "center") %>%
#   column_spec(1, bold = T) %>%         # Make the OR column bold
#   column_spec(2:4, color = "blue") %>% # Color the CI columns blue
#   scroll_box(width = "100%", height = "2000px") # Add a scroll box if the table is too large
# 
# # Print the table
# nice_table
# 
# # To display this table outside of an R Markdown document, save it to an HTML file and open it in a web browser
# save_kable(nice_table, file = "Multivariate_Results_Table_1168_subset.html")
```

The table presents the odds ratios (OR), confidence intervals (LowerCI and UpperCI), and p-values (PValue) for the variables included in the logistic regression model predicting medication adherence. Here's the interpretation of some key findings:

1. **Income Category**: The linear term (income_cat.L) has an OR of 1.17, suggesting that as income increases, there is a modest increase in the odds of medication adherence, although this is not statistically significant (p = 0.286). The quadratic term (income_cat.Q) OR of 0.85 indicates a non-linear relationship between income and adherence, but again, it is not statistically significant (p = 0.340).

2. **Insurance Classification**: The coefficients for insurance classification are represented by polynomial terms, with very high ORs for some terms (e.g., ins_classif.C OR = 14,052.26) but also very high p-values, indicating non-significant results. These might be due to overfitting or multicollinearity issues when high-order polynomial terms are included.

3. **Race/Ethnicity**: The linear term (race_6cat.L) shows a significant relationship with adherence (OR = 2.25, p = 0.021), suggesting that racial/ethnic differences do influence medication adherence.

4. **Sex**: The OR for the linear term of sex (sex.L) is 1.05, which indicates a slight increase in the odds of adherence for one sex compared to the other. However, this effect is not statistically significant (p = 0.656), suggesting that in this model, sex does not have a significant impact on medication adherence when controlling for other factors. The way sex is coded in the model, likely as a binary variable, does not indicate which sex has higher odds of adherence; it only indicates that there's no strong evidence of a difference in adherence between the sexes in this dataset.

5. **Age**: A significant relationship is seen with age, where each additional year increases the odds of adherence by about 3.4% (OR = 1.03, p < 0.00001), indicating a stronger adherence among older individuals.

6. **Education Level**: The education level variables are not significant, with ORs close to 1 and p-values well above the traditional 0.05 threshold for significance.

7. **Marital Status**: No significant effects were observed for marital status, with ORs close to 1 and non-significant p-values.

8. **Alcohol Use**: Heavy alcohol use was not significantly associated with medication adherence (OR = 1.15, p = 0.584).

9. **Tobacco Use**: The linear term for smoking history (smoking_hx.L) shows a significant negative relationship with adherence (OR = 0.70, p = 0.020), suggesting that those with a history of smoking have lower odds of adhering to medication.

It's important to note that these results should be interpreted with caution, especially for the insurance classification where the extremely large ORs and wide confidence intervals indicate potential instability in the model for these variables. This could be due to the small number of cases in certain categories or other issues with the data. Further investigation and possibly model refinement are warranted to clarify these relationships.

### Multivariate Analysis Interpretation
The multivariate analysis reveals a complex interplay of factors influencing adherence to cholesterol medication. Notably, age stands out as a significant predictor, with a slight increase in the likelihood of adherence as age increases (OR = 1.034, p < 0.0001). Tobacco use history also emerges as an important factor; those with no history of smoking are more likely to adhere to their medication regime (OR for smoking history linear term = 0.699, p = 0.02). Other factors, including income and insurance classification, though represented with polynomial terms, do not show a statistically significant effect on adherence in this model. Variables like sex, marital status, and alcohol use did not have a significant impact on medication adherence when controlling for other factors. This analysis underscores the importance of age and tobacco use in medication adherence while also suggesting that the relationships between adherence and socioeconomic factors may be nuanced and require further exploration.

## Multivariate Analysis Visualization
```{r}
# First, we need to extract the model coefficients and their confidence intervals
model_coef <- coef(summary(lr_model))
model_or <- exp(cbind(OR = model_coef[, "Estimate"], 
                      LowerCI = model_coef[, "Estimate"] - 1.96 * model_coef[, "Std. Error"], 
                      UpperCI = model_coef[, "Estimate"] + 1.96 * model_coef[, "Std. Error"]))

# Convert to a data frame for plotting
model_or_df <- as.data.frame(model_or)
model_or_df$Variable <- rownames(model_or_df)
model_or_df$OR <- exp(model_or_df$OR)  # Convert log-odds to odds

# Reorder the variables based on the OR
model_or_df <- model_or_df[order(model_or_df$OR), ]

# Plotting using ggplot2
library(ggplot2)
lr_plot <- ggplot(model_or_df, aes(x = reorder(Variable, OR), y = OR)) +
  geom_point() +
  geom_errorbar(aes(ymin = exp(LowerCI), ymax = exp(UpperCI)), width = 0.2) +
  coord_flip() +  # Flip coordinates for horizontal layout
  xlab("Variable") + ylab("Odds Ratio (OR)") +
  theme_minimal() +
  geom_hline(yintercept = 1, linetype = "dashed", color = "blue")  # Reference line for OR = 1

# Print the plot
print(lr_plot)
```

This plot is not informative as some ORs are really large and some CIs are really wide. 

## Stratified Analyses
```{r warning = FALSE}
# Stratified analysis by income level
stratified_by_income <- subset_1168_data %>%
  split(.$income_cat) %>%
  lapply(function(data) {
    glm(adherence ~ ins_classif + race_6cat + sex + age + educ_level + marital_status + alc_heavy + smoking_hx + bmi_cat + cvd_hx + alc_heavy + bmi_cat, 
        family = binomial(link = "logit"), 
        data = data)
  })

# Stratified analysis by race/ethnicity
stratified_by_race <- subset_1168_data %>%
  split(.$race_6cat) %>%
  lapply(function(data) {
    glm(adherence ~ income_cat + ins_classif + sex + age + educ_level + marital_status + alc_heavy + smoking_hx + alc_heavy + bmi_cat, 
        family = binomial(link = "logit"), 
        data = data)
  })

# Stratified analysis by gender
stratified_by_gender <- subset_1168_data %>%
  split(.$sex) %>%
  lapply(function(data) {
    glm(adherence ~ income_cat + ins_classif + race_6cat + age + educ_level + marital_status + alc_heavy + smoking_hx + alc_heavy + bmi_cat, 
        family = binomial(link = "logit"), 
        data = data)
  })

# Stratified analysis by insurance status with checks
stratified_by_insurance <- subset_1168_data %>%
  split(.$ins_classif) %>%
  lapply(function(data) {
    predictors <- c("income_cat", "race_6cat", "sex", "age", "educ_level", "marital_status", "alc_heavy", "smoking_hx")
    valid_predictors <- predictors[sapply(data[, predictors], function(x) length(unique(x)) > 1)]
    
    if(length(valid_predictors) == length(predictors)) {
      glm(adherence ~ income_cat + race_6cat + sex + age + educ_level + marital_status + alc_heavy + smoking_hx + alc_heavy + bmi_cat, 
          family = binomial(link = "logit"), 
          data = data)
    } else {
      cat("Skipped model for", names(data), "due to insufficient data.\n")
      NULL
    }
  })
stratified_by_insurance <- Filter(Negate(is.null), stratified_by_insurance)

# Stratified analysis by education level
stratified_by_education <- subset_1168_data %>%
  split(.$educ_level) %>%
  lapply(function(data) {
    glm(adherence ~ income_cat + ins_classif + race_6cat + sex + age + marital_status + alc_heavy + smoking_hx + alc_heavy + bmi_cat, 
        family = binomial(link = "logit"), 
        data = data)
  })

# Stratified analysis by marital status
stratified_by_marital <- subset_1168_data %>%
  split(.$marital_status) %>%
  lapply(function(data) {
    glm(adherence ~ income_cat + ins_classif + race_6cat + sex + age + educ_level + alc_heavy + smoking_hx + alc_heavy + bmi_cat, 
        family = binomial(link = "logit"), 
        data = data)
  })

# Stratified analysis by alcohol use
stratified_by_alcohol <- subset_1168_data %>%
  split(.$alc_heavy) %>%
  lapply(function(data) {
    glm(adherence ~ income_cat + ins_classif + race_6cat + sex + age + educ_level + marital_status + smoking_hx + alc_heavy + bmi_cat, 
        family = binomial(link = "logit"), 
        data = data)
  })

# Stratified analysis by tobacco use
stratified_by_tobacco <- subset_1168_data %>%
  split(.$smoking_hx) %>%
  lapply(function(data) {
    glm(adherence ~ income_cat + ins_classif + race_6cat + sex + age + educ_level + marital_status + alc_heavy + alc_heavy + bmi_cat, 
        family = binomial(link = "logit"), 
        data = data)
  })

# Stratified analysis by BMI
stratified_by_bmi <- subset_1168_data %>%
  split(.$bmi_cat) %>%
  lapply(function(data) {
    glm(adherence ~ income_cat + ins_classif + race_6cat + sex + age + educ_level + marital_status + alc_heavy + cvd_hx, 
        family = binomial(link = "logit"), 
        data = data)
  })

# Stratified analysis by comorbid heart disease
stratified_by_cvd <- subset_1168_data %>%
  split(.$cvd_hx) %>%
  lapply(function(data) {
    glm(adherence ~ income_cat + ins_classif + race_6cat + sex + age + educ_level + marital_status + alc_heavy + bmi_cat, 
        family = binomial(link = "logit"), 
        data = data)
  })
# Output the summaries of the stratified analyses
lapply(stratified_by_income, summary)
```

The stratified analysis by income level reveals distinct patterns of adherence to cholesterol medication across different income groups. Here's an interpretation of the results for each income category:

1. **Less than or equal to 135% of FPL (Federal Poverty Level)**:
   - **Race/Ethnicity Quadratic Component (race_6cat.Q)** shows a significant positive association with adherence, suggesting a complex, non-linear relationship between race/ethnicity and adherence in this income group.
   - **Age** shows a trend towards significance, indicating that older individuals in this income group might be more likely to adhere to medication, though this is not statistically significant at the conventional 0.05 level.
   - **Tobacco Use Linear Component (smoking_hx.L)** is significantly negatively associated with adherence, implying that individuals who have never smoked are more likely to adhere to cholesterol medication.

2. **Between 135 and 185% of FPL**:
   - **Marital Status Linear Component (marital_status.L)** shows a significant negative association with adherence, suggesting that individuals who are not married (single, divorced, widowed) are less likely to adhere to medication.
   - **Tobacco Use Linear Component (smoking_hx.L)** is again negatively associated with adherence, reinforcing the trend seen in the lower income group.

3. **Greater than 185% FPL**:
   - **Age** is significantly positively associated with adherence, indicating that older individuals in this higher income group are more likely to adhere to cholesterol medication.
   - Other variables, including marital status and tobacco use, do not show significant associations in this income group.

These findings underscore the influence of socioeconomic status on health behavior and the complex interplay of various factors like race/ethnicity, marital status, and tobacco use in influencing medication adherence. The significant variables and their directions vary across income levels, highlighting the need for targeted approaches in promoting medication adherence.

```{r}
lapply(stratified_by_race, summary)
```

The results from the stratified analysis by race/ethnicity for adherence to cholesterol medication reveal varied patterns across different groups. Here's an interpretation of the results for each subgroup:

### Mexican American
- **Significant Factors**: None of the variables show statistical significance, suggesting that within this subgroup, the included factors might not strongly predict medication adherence.
- **Trends**: There is a negative association for the linear term of sex (sex.L), indicating a potential gender-related trend in adherence, but this is not statistically significant.
- **Model Fit**: The model does not indicate a strong fit given the lack of significant predictors and a high AIC value.

### Other Hispanic
- **Significant Factors**: Marital status (marital_status.L) shows a significant negative association, suggesting that marital status might play a role in medication adherence within this subgroup.
- **Trends**: There's a marginal trend in the quadratic term of smoking history (smoking_hx.Q), hinting at a complex relationship between smoking habits and adherence, but this isn't statistically significant.
- **Model Fit**: Similar to the Mexican American group, the model does not show a strong fit due to a lack of significant predictors.

### Non-Hispanic White
- **Significant Factors**: Income (income_cat.Q) and age show significant negative and positive associations, respectively, indicating these are important predictors of adherence in this subgroup.
- **Trends**: There are marginal trends in smoking history, both linear and quadratic terms, suggesting a nuanced relationship between smoking and adherence.
- **Model Fit**: The presence of significant predictors implies a better model fit compared to the previous groups.

### Non-Hispanic Black
- **Significant Factors**: Income (income_cat.L) and sex (sex.L) are significant predictors, indicating that these factors are influential in medication adherence within this subgroup.
- **Trends**: There's a marginal negative trend in the linear term of smoking history.
- **Model Fit**: The model shows a better fit with significant predictors compared to other groups.

### Non-Hispanic Asian
- **Significant Factors**: Income (income_cat.L) and age are significant predictors of adherence, with age showing a positive association.
- **Trends**: The model couldn't estimate the association for alcohol use due to data limitations (indicated by NA).
- **Model Fit**: The significant predictors indicate a better model fit, although some coefficients could not be estimated.

### Other Race (including multiracial)
- **Significant Factors**: No significant predictors are found in this group.
- **Trends**: Large standard errors and non-significant p-values suggest that the model may not be well specified for this group, or there may be insufficient data.
- **Model Fit**: The lack of significant predictors and the presence of large standard errors suggest a poor model fit.

These results underscore the importance of considering race/ethnicity in understanding medication adherence patterns, as different factors appear to be influential in different subgroups. The variability in significance and trends across these groups highlights the complexity of factors influencing medication adherence and the necessity of tailored approaches in different racial/ethnic contexts.

```{r}
lapply(stratified_by_gender, summary)
```

The results from the stratified analysis by gender for adherence to cholesterol medication present distinct patterns between males and females. Here's an interpretation of the results for each gender:

### Male
- **Significant Factors**: Race (race_6cat.Q) and age are significant predictors, with age showing a strong positive association with adherence.
- **Trends**: The model suggests a marginal negative trend in sex (sex.L), indicating potential differences in adherence based on gender, but this is not statistically significant.
- **Model Fit**: The presence of significant predictors suggests a reasonable model fit, although several predictors are not significant, indicating a complex relationship between these factors and adherence.

### Female
- **Significant Factors**: Race (race_6cat.L) and marital status (marital_status.L) show significant associations with adherence. The positive association with race suggests certain racial/ethnic groups may have higher adherence rates. Negative association with marital status indicates that certain marital statuses might be associated with lower adherence.
- **Trends**: Age is also a significant positive predictor, similar to males. The model does not show significant associations for other factors like income, insurance, and educational level.
- **Model Fit**: The model shows a reasonable fit, with significant predictors indicating a better understanding of adherence patterns among females.

In summary, the stratified analysis by gender reveals that certain factors like race and age play a significant role in predicting adherence to cholesterol medication, but this varies between males and females. The results highlight the need for gender-specific approaches in understanding and improving medication adherence. The lack of significant associations for some variables suggests the complexity of factors influencing adherence and the possibility of other unmeasured variables playing a role.

```{r}
lapply(stratified_by_insurance, summary)
```

The results from the stratified analysis by insurance type for adherence to cholesterol medication show varied findings across different insurance categories. Here's an interpretation of the results for each category:

### Both Private and Government, with Drug Coverage
- **Significant Factors**: None of the factors are statistically significant in predicting adherence.
- **Model Fit**: The model does not show a strong association between the included factors and medication adherence for this group. The high AIC value and large standard errors suggest the model may not be well-fitted to the data.

### Both Private and Government, without Drug Coverage
- **Model Issues**: This model presents extremely large standard errors and estimates, which suggests issues with the data or model fit. The results are not meaningful for interpretation.
- **Data Limitations**: The results indicate possible limitations in the data, such as small sample size or lack of variation in predictor variables for this subgroup.

### Government, with Drug Coverage
- **Significant Factors**: Race (race_6cat.L) and age are significant predictors. A positive coefficient for age indicates higher likelihood of adherence as age increases.
- **Model Fit**: The model seems better fitted for this group compared to the previous categories, as indicated by more reasonable standard errors and significant predictors.

### Government, without Drug Coverage
- **Model Issues**: Like the model for "Both private and government, without drug coverage," this model also shows extremely large standard errors and estimates, suggesting issues with the data or model fit for this group.
- **Data Limitations**: This result suggests potential data limitations or insufficient sample size for meaningful analysis in this subgroup.

### Private, with Drug Coverage
- **Trends**: No significant predictors were found. However, there's a marginal negative trend with education level (educ_level.L) and a positive trend with age.
- **Model Fit**: The model has better fit compared to some other categories but lacks strong predictive power.

### Private, without Drug Coverage
- **Model Issues**: This model faces similar issues as seen in the "Government, without drug coverage" and "Both private and government, without drug coverage" categories, with extremely large estimates and standard errors, rendering the results unreliable.

In summary, the stratified analysis by insurance type reveals considerable variation in the model's predictive power across different insurance groups. While age and race emerge as significant predictors in some categories, the overall effectiveness of the model varies, with several categories showing potential data limitations or issues with model fit. This variation highlights the complexity of the relationship between insurance type and medication adherence, necessitating a nuanced approach for each category.

```{r}
lapply(stratified_by_education, summary)
```

The stratified analysis by education level for adherence to cholesterol medication reveals varying influences of demographic and socioeconomic factors across different educational groups. Here's an interpretation for each group:

1. **Less than 9th grade**: In this group, age was the only variable with a trend towards significance (p = 0.116), suggesting a potential increase in adherence with age. However, none of the variables reached statistical significance, indicating that factors like income, insurance classification, race, sex, marital status, alcohol, and tobacco use did not have a strong, detectable influence on medication adherence in this subgroup.

2. **9th-11th grade**: Here, income (linear term) showed a trend towards significance (p = 0.0835), and age was significantly associated with adherence (p = 0.0353). This suggests that individuals in higher income categories and older ages may be more likely to adhere to medication in this educational group.

3. **High School graduate or GED**: Tobacco use was significantly associated with non-adherence (p = 0.00678), indicating that smokers in this educational group are less likely to adhere to medication. Other factors, including income, insurance, race, sex, marital status, and alcohol use, did not show significant associations.

4. **Some college or AA degree**: Race (linear term) was significantly associated with adherence (p = 0.00646), indicating that racial factors play a role in medication adherence in this group. Other factors like income, insurance, sex, age, marital status, alcohol, and tobacco use did not show a strong association.

5. **College graduate or above**: Age (p = 0.00746) and race (cubic term, p = 0.08727) were significantly associated with adherence, with older individuals and certain racial groups being more likely to adhere to medication. Other variables, including income, insurance, sex, marital status, alcohol, and tobacco use, did not exhibit a significant influence on adherence.

Overall, these results indicate that the factors influencing medication adherence can vary considerably across different education levels. While age consistently appears as a significant or near-significant factor across most groups, the influence of other variables like income, insurance status, race, and health behaviors (smoking and alcohol use) varies. This highlights the complexity of medication adherence as a behavior influenced by a combination of demographic, socioeconomic, and lifestyle factors.

```{r}
lapply(stratified_by_marital, summary)
```

The stratified analysis by marital status for adherence to cholesterol medication demonstrates that the influence of various factors differs across marital status categories. Here's an interpretation for each group:

1. **Married/Living with Partner**: In this group, age was a significant predictor (p = 1.24e-05), indicating higher adherence with increasing age. Additionally, the quadratic term of income category showed significance (p = 0.0122), suggesting a non-linear relationship between income and adherence. Alcohol use (p = 0.0794) and smoking history (linear term, p = 0.0522) showed trends towards significance, indicating possible influences on adherence.

2. **Widowed/Divorced/Separated**: None of the variables reached statistical significance in this group. This suggests that factors like income, insurance classification, race, sex, age, education level, alcohol, and tobacco use did not have a strong, detectable influence on medication adherence among individuals who are widowed, divorced, or separated.

3. **Never Married**: In this subgroup, the quadratic term of income category showed a trend towards significance (p = 0.0563), hinting at a possible complex relationship between income levels and medication adherence. Other factors, including insurance classification, race, sex, age, education level, alcohol, and tobacco use, did not exhibit a significant influence on adherence.

Overall, these results indicate that marital status can influence how different socioeconomic and demographic factors affect medication adherence. Age consistently appears as a significant predictor in the married/living with partner group, while income shows a complex relationship in the same group and a trend towards significance in the never married group. This highlights the importance of considering marital status when examining factors influencing medication adherence.

```{r}
lapply(stratified_by_alcohol, summary)
```

The stratified analysis by alcohol use for adherence to cholesterol medication yields distinct insights for individuals who do not use alcohol (`FALSE`) compared to those who do (`TRUE`).

1. **For Non-Alcohol Users (`FALSE`)**:
   - Age is a significant predictor (p = 0.000616), indicating higher adherence with increasing age.
   - Race, specifically the linear term of race category (p = 0.038033), is significant, suggesting certain racial/ethnic groups might have different adherence patterns.
   - Smoking history, particularly the linear term (p = 0.044860), is significant, hinting at a potential relationship between smoking behavior and medication adherence.
   - Other factors, such as income, insurance classification, sex, education level, and marital status, did not show significant associations with medication adherence.

2. **For Alcohol Users (`TRUE`)**:
   - Age again emerges as a significant factor (p = 0.0026), reinforcing the trend seen in non-alcohol users.
   - Marital status, particularly the quadratic term, is significant (p = 0.0354), suggesting that marital status might play a different role in medication adherence for alcohol users.
   - Race, sex, education level, and other factors did not show significant associations.
   - Smoking history showed a trend towards significance (p = 0.0656 for the linear term), indicating a potential relationship between smoking and adherence in this subgroup.

Overall, these results highlight that alcohol use can modify the relationship between various demographic and socioeconomic factors and medication adherence. Age consistently appears as a significant predictor in both groups, while race and smoking history show significance in the non-alcohol user group. This suggests that interventions to improve medication adherence may need to be tailored differently for individuals based on their alcohol use status.

```{r}
lapply(stratified_by_tobacco, summary)
```

The stratified analysis by tobacco use (Never, Past, Current) for adherence to cholesterol medication provides varied results across different groups:

1. **Never Users**:
   - Age is the only significant predictor (p = 0.00027), indicating higher adherence with increasing age.
   - Other variables, including income, insurance classification, race, sex, education level, marital status, and alcohol use, did not show significant associations with adherence.

2. **Past Users**:
   - No variables showed statistical significance in predicting medication adherence.
   - The model includes income, insurance classification, race, sex, age, education level, marital status, and alcohol use, but none demonstrated a clear impact on adherence.

3. **Current Users**:
   - Age and race emerge as significant predictors. Age is significant (p = 0.0076), reinforcing the trend seen in never users.
   - The linear term of race (race_6cat.C) is significant (p = 0.0295), suggesting that certain racial/ethnic groups among current tobacco users might have different adherence patterns.
   - Other factors, including income, insurance classification, sex, education level, marital status, and alcohol use, did not show significant associations with adherence.

These results illustrate that the factors influencing medication adherence may vary based on tobacco use history. Age consistently appears as a significant factor across never and current users. The impact of race is notably different in current users compared to never and past users. This suggests that interventions to improve medication adherence could benefit from considering an individual's tobacco use history.

### Stratified Analysis Interpretation
The stratified analyses for adherence to cholesterol medication across various demographic and behavioral groups revealed nuanced insights. Age consistently emerged as a significant factor, particularly among non-tobacco users and current tobacco users, indicating higher adherence with increasing age. Marital status showed significance in certain strata, hinting at the influence of social support systems on adherence. Notably, the influence of race varied across different tobacco usage groups, suggesting diverse adherence patterns among racial groups, especially among current tobacco users. Income, insurance status, and other socioeconomic factors, while included in the models, did not consistently show a significant impact across different strata. This highlights the complexity of medication adherence behavior, which appears to be influenced by a combination of demographic, socioeconomic, and personal behavior factors.


```{r}
lapply(stratified_by_bmi, summary)
```

```{r}
lapply(stratified_by_cvd, summary)
```

## Summary: 
The comprehensive study on adherence to cholesterol medication encompassed multiple analytical approaches, revealing intricate patterns influenced by a blend of demographic, socioeconomic, and behavioral factors:

1. **Descriptive Analysis**: This provided a basic understanding of the study population, highlighting key characteristics such as age distribution, gender, income levels, education, marital status, and tobacco and alcohol use. These factors set the foundation for more detailed analyses.

2. **Bivariate Analysis**: This analysis revealed significant associations between medication adherence and various factors. Age emerged as a key determinant, with older individuals showing higher adherence. Marital status also appeared influential, suggesting the role of social support in medication compliance. However, income, insurance, and race/ethnicity associations with adherence were less clear, indicating a need for more nuanced analysis.

3. **Multivariate Analysis**: The logistic regression model integrated various predictors, providing a more comprehensive view. Age maintained its significance, reaffirming its role in adherence. Gender, income, and insurance classifications showed some influence, but their impact was less pronounced than age. Behavioral factors like tobacco use also emerged as significant, pointing to broader health-conscious behaviors among non-users.

4. **Stratified Analysis**: This approach delved deeper into specific subgroups, uncovering distinct patterns within each. Age consistently showed significance across various strata, reinforcing its overall importance. Marital status and race displayed varying levels of influence depending on the stratification factor, such as tobacco use, suggesting complex interactions between these variables and adherence behavior.

Overall, the study findings underscore the multifaceted nature of medication adherence. Age consistently stands out as a crucial factor, with the support system provided by marital status and the lifestyle choices indicated by tobacco use also playing significant roles. The varying influence of socioeconomic factors across different analyses suggests that these elements interact with other variables in complex ways, influencing adherence. These insights are critical for designing targeted interventions and policies to improve medication adherence and address health disparities in cholesterol management.
